requirements.txt
python-socketio
eventlet
argon2-cffi